==> DOGE/01.Data_Fetch.ipynb <==
# imports
import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
import seaborn

from binance.client import Client
from datetime import datetime, timezone

import time
import csv
import gc
import os

#  client
API_Key = np.load('ReadOnlykeys.npy')[0]
Secret_Key = np.load('ReadOnlykeys.npy')[1]
client = Client(api_key=API_Key, api_secret=Secret_Key)

# get historical data

# Convert ISO 8601 date strings to Unix timestamp (milliseconds)
def iso_to_unix(iso_str):
    dt = datetime.strptime(iso_str, "%Y-%m-%dT%H:%M:%SZ").replace(tzinfo=timezone.utc)
    return int(dt.timestamp() * 1000)  # Convert to milliseconds

# Define the time intervals
start_time = iso_to_unix("2020-09-01T00:00:00Z")
end_time = iso_to_unix("2024-01-01T00:00:00Z")
# end_time = iso_to_unix("2023-10-01T00:00:00Z")


# Open a CSV file for writing
# save all result into csv
csv_path = "Data/historical_DOGEUSDT_3min_data.csv"
if not os.path.exists(csv_path): # Write header
    index =["Open time", "Open", "High", "Low", "Close", "Volume", "Close time", "Quote asset volume", "Number of trades", "Taker buy base asset volume", "Taker buy quote asset volume", "Ignore"]
    with open(csv_path, mode='a', newline='') as file:
        writer = csv.writer(file)
        writer.writerow(index)

# Fetch the data in chunks
while start_time < end_time:
    time.sleep(0.125)
    chunk = client.get_historical_klines(
        symbol="DOGEUSDT",
        interval=Client.KLINE_INTERVAL_3MINUTE,
        start_str=start_time,
        end_str=end_time,
        limit=1000  # max limit per request
    )
    if not chunk:
        break  # Break the loop if no more data is returned

    # Write the chunk to the CSV file
    with open(csv_path, mode='a', newline='') as file:
        writer = csv.writer(file)
        writer.writerows(chunk)
        file.flush()

    # Update the start_time to the timestamp of the last kline in the chunk
    start_time = chunk[-1][0] + 15 * 60 * 1000  # Add 15 minutes in milliseconds

    # Optional: Force garbage collection to free up memory
    del chunk
    gc.collect()

# takes around 5 minutes

# get 2023 data

# Convert ISO 8601 date strings to Unix timestamp (milliseconds)
def iso_to_unix(iso_str):
    dt = datetime.strptime(iso_str, "%Y-%m-%dT%H:%M:%SZ").replace(tzinfo=timezone.utc)
    return int(dt.timestamp() * 1000)  # Convert to milliseconds

# Define the time intervals
start_time = iso_to_unix("2023-01-01T00:00:00Z")
end_time = iso_to_unix("2024-01-01T00:00:00Z")
# end_time = iso_to_unix("2023-10-01T00:00:00Z")


# Open a CSV file for writing
# save all result into csv
csv_path = "Data/2023_DOGEUSDT_3min_data.csv"
if not os.path.exists(csv_path): # Write header
    index =["Open time", "Open", "High", "Low", "Close", "Volume", "Close time", "Quote asset volume", "Number of trades", "Taker buy base asset volume", "Taker buy quote asset volume", "Ignore"]
    with open(csv_path, mode='a', newline='') as file:
        writer = csv.writer(file)
        writer.writerow(index)

# Fetch the data in chunks
while start_time < end_time:
    time.sleep(0.125)
    chunk = client.get_historical_klines(
        symbol="DOGEUSDT",
        interval=Client.KLINE_INTERVAL_3MINUTE,
        start_str=start_time,
        end_str=end_time,
        limit=1000  # max limit per request
    )
    if not chunk:
        break  # Break the loop if no more data is returned

    # Write the chunk to the CSV file
    with open(csv_path, mode='a', newline='') as file:
        writer = csv.writer(file)
        writer.writerows(chunk)
        file.flush()

    # Update the start_time to the timestamp of the last kline in the chunk
    start_time = chunk[-1][0] + 15 * 60 * 1000  # Add 15 minutes in milliseconds

    # Optional: Force garbage collection to free up memory
    del chunk
    gc.collect()

# takes around 2 minutes

# get 2024 data

# Convert ISO 8601 date strings to Unix timestamp (milliseconds)
def iso_to_unix(iso_str):
    dt = datetime.strptime(iso_str, "%Y-%m-%dT%H:%M:%SZ").replace(tzinfo=timezone.utc)
    return int(dt.timestamp() * 1000)  # Convert to milliseconds

# Define the time intervals
start_time = iso_to_unix("2024-01-01T00:00:00Z")
end_time = iso_to_unix("2024-02-01T00:00:00Z")
# end_time = iso_to_unix("2023-10-01T00:00:00Z")


# Open a CSV file for writing
# save all result into csv
csv_path = "Data/2024_DOGEUSDT_3min_data.csv"
if not os.path.exists(csv_path): # Write header
    index =["Open time", "Open", "High", "Low", "Close", "Volume", "Close time", "Quote asset volume", "Number of trades", "Taker buy base asset volume", "Taker buy quote asset volume", "Ignore"]
    with open(csv_path, mode='a', newline='') as file:
        writer = csv.writer(file)
        writer.writerow(index)

# Fetch the data in chunks
while start_time < end_time:
    time.sleep(0.125)
    chunk = client.get_historical_klines(
        symbol="DOGEUSDT",
        interval=Client.KLINE_INTERVAL_3MINUTE,
        start_str=start_time,
        end_str=end_time,
        limit=1000  # max limit per request
    )
    if not chunk:
        break  # Break the loop if no more data is returned

    # Write the chunk to the CSV file
    with open(csv_path, mode='a', newline='') as file:
        writer = csv.writer(file)
        writer.writerows(chunk)
        file.flush()

    # Update the start_time to the timestamp of the last kline in the chunk
    start_time = chunk[-1][0] + 15 * 60 * 1000  # Add 15 minutes in milliseconds

    # Optional: Force garbage collection to free up memory
    del chunk
    gc.collect()

# takes around 1 minutes
